Moved visibility property to default controller.
